Altered insert scada to actually castrate this program. Changed scan and passing of values to double instead of floats.

+++ scadaTransfer/src/scadaTransfer.h
@@ -24,7 +24,7 @@ int insertScada(
 #ifdef ANSIPROTO
 SQL_DATE *,
 int,
-float *,
+double *,
 ID,
 char *
 #endif
